[fs] nfsd: let nfsd_symlink assume null-terminated data
Message-id: <1410898849-22126-16-git-send-email-bfields@redhat.com>
Patchwork-id: 95698
O-Subject: [Patch RHEL7 015/165] nfsd: let nfsd_symlink assume null-terminated data
Bugzilla: 1044798 1123721 1141932
RH-Acked-by: Benjamin Coddington <bcodding@redhat.com>
RH-Acked-by: Steve Dickson <SteveD@redhat.com>

Currently nfsd_symlink has a weird hack to serve callers who don't
null-terminate symlink data: it looks ahead at the next byte to see if
it's zero, and copies it to a new buffer to null-terminate if not.

That means callers don't have to null-terminate, but they *do* have to
ensure that the byte following the end of the data is theirs to read.

That's a bit subtle, and the NFSv4 code actually got this wrong.

So let's just throw out that code and let callers pass null-terminated
strings; we've already fixed them to do that.

Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Upstream: 52ee04330f585d1b5bc40442f07df07248fa3aee
Bugzilla: 1044798
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>