[fs] nfsd: add a new /proc/fs/nfsd/max_connections file
Message-id: <1410898849-22126-27-git-send-email-bfields@redhat.com>
Patchwork-id: 95758
O-Subject: [Patch RHEL7 026/165] nfsd: add a new /proc/fs/nfsd/max_connections file
Bugzilla: 1044798 1123721 1141932
RH-Acked-by: Benjamin Coddington <bcodding@redhat.com>
RH-Acked-by: Steve Dickson <SteveD@redhat.com>

From: Jeff Layton <jlayton@primarydata.com>

Currently, the maximum number of connections that nfsd will allow
is based on the number of threads spawned. While this is fine for a
default, there really isn't a clear relationship between the two.

The number of threads corresponds to the number of concurrent requests
that we want to allow the server to process at any given time. The
connection limit corresponds to the maximum number of clients that we
want to allow the server to handle. These are two entirely different
quantities.

Break the dependency on increasing threads in order to allow for more
connections, by adding a new per-net parameter that can be set to a
non-zero value. The default is still to base it on the number of threads,
so there should be no behavior change for anyone who doesn't use it.

Cc: Trond Myklebust <trond.myklebust@primarydata.com>
Signed-off-by: Jeff Layton <jlayton@primarydata.com>
Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Upstream: 5b8db00bae39e5ecd9bafb05478cca5b42564ab1
Bugzilla: 1044798
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>