[fs] nfsd: Allow struct nfsd4_compound_state to cache the nfs4_client
Message-id: <1410898849-22126-40-git-send-email-bfields@redhat.com>
Patchwork-id: 95719
O-Subject: [Patch RHEL7 039/165] nfsd: Allow struct nfsd4_compound_state to cache the nfs4_client
Bugzilla: 1044798 1123721 1141932
RH-Acked-by: Benjamin Coddington <bcodding@redhat.com>
RH-Acked-by: Steve Dickson <SteveD@redhat.com>

From: Jeff Layton <jlayton@primarydata.com>

We want to use the nfsd4_compound_state to cache the nfs4_client in
order to optimise away extra lookups of the clid.

In the v4.0 case, we use this to ensure that we only have to look up the
client at most once per compound for each call into lookup_clientid. For
v4.1+ we set the pointer in the cstate during SEQUENCE processing so we
should never need to do a search for it.

Signed-off-by: Trond Myklebust <trond.myklebust@primarydata.com>
Reviewed-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Upstream: 4b24ca7d30430882a2eaeb9d511990fb4581230d
Bugzilla: 1044798
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>