[fs] nfsd: set stateid access and deny bits in nfs4_get_vfs_file
Message-id: <1410898849-22126-53-git-send-email-bfields@redhat.com>
Patchwork-id: 95726
O-Subject: [Patch RHEL7 052/165] nfsd: set stateid access and deny bits in nfs4_get_vfs_file
Bugzilla: 1044798 1123721 1141932
RH-Acked-by: Benjamin Coddington <bcodding@redhat.com>
RH-Acked-by: Steve Dickson <SteveD@redhat.com>

From: Jeff Layton <jlayton@primarydata.com>

Cleanup -- ensure that the stateid bits are set at the same time that
the file access refcounts are incremented. Keeping them coherent like
this makes it easier to ensure that we account for all of the
references.

Since the initialization of the st_*_bmap fields is done when it's
hashed, we go ahead and hash the stateid before getting access to the
file and unhash it if that function returns error. This will be
necessary anyway in a follow-on patch that will overhaul deny mode
handling.

Signed-off-by: Jeff Layton <jlayton@primarydata.com>
Reviewed-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Upstream: 6eb3a1d096751bcdec8fd9d9bb565fa9cba5897f
Bugzilla: 1044798
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>