[fs] nfsd: Convert delegation counter to an atomic_long_t type
Message-id: <1410898849-22126-80-git-send-email-bfields@redhat.com>
Patchwork-id: 95730
O-Subject: [Patch RHEL7 079/165] nfsd: Convert delegation counter to an atomic_long_t type
Bugzilla: 1044798 1123721 1141932
RH-Acked-by: Benjamin Coddington <bcodding@redhat.com>
RH-Acked-by: Steve Dickson <SteveD@redhat.com>

From: Trond Myklebust <trond.myklebust@primarydata.com>

We want to convert to an atomic type so that we don't need to lock
across the call to alloc_init_deleg(). Then convert to a long type so
that we match the size of 'max_delegations'.

None of this is a problem today, but it will be once we remove
client_mutex protection.

Signed-off-by: Trond Myklebust <trond.myklebust@primarydata.com>
Reviewed-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Upstream: 02a3508dba9a58b7bd77cc91f8e941e2dda94d1d
Bugzilla: 1044798
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>