[fs] nfsd: clean up lockowner refcounting when finding them
Message-id: <1410898849-22126-110-git-send-email-bfields@redhat.com>
Patchwork-id: 95779
O-Subject: [Patch RHEL7 109/165] nfsd: clean up lockowner refcounting when finding them
Bugzilla: 1044798 1123721 1141932
RH-Acked-by: Benjamin Coddington <bcodding@redhat.com>
RH-Acked-by: Steve Dickson <SteveD@redhat.com>

From: Jeff Layton <jlayton@primarydata.com>

Ensure that when finding or creating a lockowner, that we get a
reference to it. For now, we also take an extra reference when a
lockowner is created that can be put when release_lockowner is called,
but we'll remove that in a later patch once we change how references are
held.

Since we no longer destroy lockowners in the event of an error in
nfsd4_lock, we must change how the seqid gets bumped in the lk_is_new
case. Instead of doing so on creation, do it manually in nfsd4_lock.

Signed-off-by: Jeff Layton <jlayton@primarydata.com>
Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Upstream: 5db1c03feb00a72846927172d81744790e601b97
Bugzilla: 1044798
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>