[fs] nfsd: make openstateids hold references to their openowners
Message-id: <1410898849-22126-114-git-send-email-bfields@redhat.com>
Patchwork-id: 95729
O-Subject: [Patch RHEL7 113/165] nfsd: make openstateids hold references to their openowners
Bugzilla: 1044798 1123721 1141932
RH-Acked-by: Benjamin Coddington <bcodding@redhat.com>
RH-Acked-by: Steve Dickson <SteveD@redhat.com>

From: Jeff Layton <jlayton@primarydata.com>

Change it so that only openstateids hold persistent references to
openowners. References can still be held by compounds in progress.

With this, we can get rid of NFS4_OO_NEW. It's possible that we
will create a new openowner in the process of doing the open, but
something later fails. In the meantime, another task could find
that openowner and start using it on a successful open. If that
occurs we don't necessarily want to tear it down, just put the
reference that the failing compound holds.

Signed-off-by: Jeff Layton <jlayton@primarydata.com>
Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Upstream: d3134b1049c3db8fdac0c6dc9430facf30015a15
Bugzilla: 1044798
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>