[fs] nfsd: don't thrash the cl_lock while freeing an open stateid
Message-id: <1410898849-22126-124-git-send-email-bfields@redhat.com>
Patchwork-id: 95747
O-Subject: [Patch RHEL7 123/165] nfsd: don't thrash the cl_lock while freeing an open stateid
Bugzilla: 1044798 1123721 1141932
RH-Acked-by: Benjamin Coddington <bcodding@redhat.com>
RH-Acked-by: Steve Dickson <SteveD@redhat.com>

From: Jeff Layton <jlayton@primarydata.com>

When we remove the client_mutex, we'll have a potential race between
FREE_STATEID and CLOSE.

The root of the problem is that we are walking the st_locks list,
dropping the spinlock and then trying to release the persistent
reference to the lockstateid. In between, a FREE_STATEID call can come
along and take the lock, find the stateid and then try to put the
reference. That leads to a double put.

Fix this by not releasing the cl_lock in order to release each lock
stateid. Use put_generic_stateid_locked to unhash them and gather them
onto a list, and free_ol_stateid_reaplist to free any that end up on the
list.

Signed-off-by: Jeff Layton <jlayton@primarydata.com>
Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Upstream: d83017f94c290c56010e194cdbc45e59894ccae2
Bugzilla: 1044798
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>