[fs] nfsd: don't destroy clients that are busy
Message-id: <1410898849-22126-141-git-send-email-bfields@redhat.com>
Patchwork-id: 95750
O-Subject: [Patch RHEL7 140/165] nfsd: don't destroy clients that are busy
Bugzilla: 1044798 1123721 1141932
RH-Acked-by: Benjamin Coddington <bcodding@redhat.com>
RH-Acked-by: Steve Dickson <SteveD@redhat.com>

From: Jeff Layton <jlayton@primarydata.com>

It's possible that we'll have an in-progress call on some of the clients
while a rogue EXCHANGE_ID or DESTROY_CLIENTID call comes in. Be sure to
try and mark the client expired first, so that the refcount is
respected.

This will only be a problem once the client_mutex is removed.

Signed-off-by: Jeff Layton <jlayton@primarydata.com>
Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Upstream: fd699b8a48c0ca36e782cf705794358b3e4b8c25
Bugzilla: 1044798
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>