[fs] nfsd: protect clid and verifier generation with client_lock
Message-id: <1410898849-22126-142-git-send-email-bfields@redhat.com>
Patchwork-id: 95713
O-Subject: [Patch RHEL7 141/165] nfsd: protect clid and verifier generation with client_lock
Bugzilla: 1044798 1123721 1141932
RH-Acked-by: Benjamin Coddington <bcodding@redhat.com>
RH-Acked-by: Steve Dickson <SteveD@redhat.com>

From: Jeff Layton <jlayton@primarydata.com>

The clid counter is a global counter currently. Move it to be a per-net
property so that it can be properly protected by the nn->client_lock
instead of relying on the client_mutex.

The verifier generator is also potentially racy if there are two
simultaneous callers. Generate the verifier when we generate the clid
value, so it's also created under the client_lock. With this, there's
no need to keep two counters as they'd always be in sync anyway, so
just use the clientid_counter for both.

As Trond points out, what would be best is to eventually move this
code to use IDR instead of the hash tables. That would also help ensure
uniqueness, but that's probably best done as a separate project.

Signed-off-by: Jeff Layton <jlayton@primarydata.com>
Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Upstream: 294ac32e99861f6efee548a6b7afb27c32cd502f
Bugzilla: 1044798
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>