[fs] nfs4: copy acceptor name from context to nfs_client
Message-id: <1411050057-12417-27-git-send-email-steved@redhat.com>
Patchwork-id: 96217
O-Subject: [RHEL7.1 PATCH 26/94] nfs4: copy acceptor name from context to nfs_client
Bugzilla: 1111708
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>
RH-Acked-by: Benjamin Coddington <bcodding@redhat.com>

From: Jeff Layton <jlayton@poochiereds.net>

The current CB_COMPOUND handling code tries to compare the principal
name of the request with the cl_hostname in the client. This is not
guaranteed to ever work, particularly if the client happened to mount
a CNAME of the server or a non-fqdn.

Fix this by instead comparing the cr_principal string with the acceptor
name that we get from gssd. In the event that gssd didn't send one
down (i.e. it was too old), then we fall back to trying to use the
cl_hostname as we do today.

Signed-off-by: Jeff Layton <jlayton@poochiereds.net>
Signed-off-by: Trond Myklebust <trond.myklebust@primarydata.com>
Upstream Id: f11b2a1cfbf5dd783eb55cb470509d06e20d1c78
Bugzilla: 1111708
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>