[crypto] pkcs7: Provide a single place to do signed info block freeing
Message-id: <1411414897-26220-4-git-send-email-vgoyal@redhat.com>
Patchwork-id: 96446
O-Subject: [RHEL7.1 PATCH 3/4] PKCS#7: Provide a single place to do signed info block freeing
Bugzilla: 1129779
RH-Acked-by: David Howells <dhowells@redhat.com>

BZ:1129779

This commit id is in linux-security tree of James Morris.

commit 3cd0920cde6a726b9965b60184f9d5a54358304d
Author: David Howells <dhowells@redhat.com>
Date: Tue, 16 Sep 2014 17:29:03 +0100

The code to free a signed info block is repeated several times, so move the
code to do it into a function of its own.  This gives us a place to add clean
ups for stuff that gets added to pkcs7_signed_info.

Signed-off-by: David Howells <dhowells@redhat.com>
Acked-by: Vivek Goyal <vgoyal@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>