[crypto] pkcs7: Fix the parser cleanup to drain parsed out X.509 certs
Message-id: <1411414897-26220-5-git-send-email-vgoyal@redhat.com>
Patchwork-id: 96449
O-Subject: [RHEL7.1 PATCH 4/4] PKCS#7: Fix the parser cleanup to drain parsed out X.509 certs
Bugzilla: 1129779
RH-Acked-by: David Howells <dhowells@redhat.com>

BZ:1129779

This commit id is in linux-security tree of James Morris.

commit cecf5d2e1208da512a4c951c24acd66c54a4d06c
Author: David Howells <dhowells@redhat.com>
Date: Tue, 16 Sep 2014 17:29:03 +0100

Fix the parser cleanup code to drain parsed out X.509 certs in the case that
the decode fails and we jump to error_decode.

The function is rearranged so that the same cleanup code is used in the success
case as the error case - just that the message descriptor under construction is
only released if it is still pointed to by the context struct at that point.

Signed-off-by: David Howells <dhowells@redhat.com>
Acked-by: Vivek Goyal <vgoyal@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>