[redhat] configs: Enable NFSoRDMA server
Message-id: <cb113c6d379170d0e79aa12629b2bb32865eb36f.1411578494.git.dledford@redhat.com>
Patchwork-id: 96661
O-Subject: [Patch RHEL7 12/26] Enable NFSoRDMA server
Bugzilla: 1107621
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Steve Dickson <SteveD@redhat.com>

We disabled this right before rhel7.0 went out due to a hard to track
down bug and lack of confidence in the overall code.  Since then, a
considerable amount of work has happened upstream, and been backported
to rhel7.1.  I think we can enable this now, and if we have any doubts,
then mark it as tech preview.

Signed-off-by: Doug Ledford <dledford@redhat.com>
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>