[infiniband] ipoib: Make the carrier_on_task race aware
Message-id: <740daa28a83126cecda945aa061b2ae6061e4035.1411578494.git.dledford@redhat.com>
Patchwork-id: 96674
O-Subject: [Patch RHEL7 19/26] IPoIB: Make the carrier_on_task race aware
Bugzilla: 1107621
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Steve Dickson <SteveD@redhat.com>

We blindly assume that we can just take the rtnl lock and that will
prevent races with downing this interface.  Unfortunately, that's not
the case.  In ipoib_mcast_stop_thread() we will call flush_workqueue()
in an attempt to clear out all remaining instances of ipoib_join_task.
But, since this task is put on the same workqueue as the join task, the
flush_workqueue waits on this thread too.  But this thread is deadlocked
on the rtnl lock.  The better thing here is to use trylock and loop on
that until we either get the lock or we see that FLAG_ADMIN_UP has
been cleared, in which case we don't need to do anything anyway and we
just return.

Signed-off-by: Doug Ledford <dledford@redhat.com>
(cherry picked from commit a1a3e7c3ef3508c6ae0b5e62eddf15ecc809c2a3)
Signed-off-by: Doug Ledford <dledford@redhat.com>
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>