[infiniband] ipoib: Make ipoib_mcast_stop_thread flush the workqueue
Message-id: <288895bab031a961176f324029e58e17eae2c217.1411578494.git.dledford@redhat.com>
Patchwork-id: 96680
O-Subject: [Patch RHEL7 24/26] IPoIB: Make ipoib_mcast_stop_thread flush the workqueue
Bugzilla: 1107621
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Steve Dickson <SteveD@redhat.com>

We used to pass a flush variable to mcast_stop_thread to indicate if we
should flush the workqueue or not.  This was due to some code trying to
flush a workqueue that it was currently running on which is a no-no.
Now that we have per-device work queues, and now that
ipoib_mcast_restart_task has taken the fact that it is queued on a
single thread workqueue with all of the ipoib_mcast_join_task's and
therefore has no need to stop the join task while it runs, we can do
away with the flush parameter and unilaterally flush always.

Signed-off-by: Doug Ledford <dledford@redhat.com>
(cherry picked from commit 82a299931fba5c012b47e8947d98d01c3accf8f3)
Signed-off-by: Doug Ledford <dledford@redhat.com>
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>