[virt] hyperv: Enable sendbuf mechanism on the send path
Message-id: <1411952358-34104-23-git-send-email-jasowang@redhat.com>
Patchwork-id: 96762
O-Subject: [RHEL7.1 PATCH 22/34] hyperv: Enable sendbuf mechanism on the send path
Bugzilla: 1146357
RH-Acked-by: Radim Krcmar <rkrcmar@redhat.com>
RH-Acked-by: Vitaly Kuznetsov <vkuznets@redhat.com>

From: KY Srinivasan <kys@microsoft.com>

Bugzilla: 1146357

We send packets using a copy-free mechanism (this is the Guest to Host transport
via VMBUS). While this is obviously optimal for large packets,
it may not be optimal for small packets. Hyper-V host supports
a second mechanism for sending packets that is "copy based". We implement that
mechanism in this patch.

In this version of the patch I have addressed a comment from David Miller.

With this patch (and all of the other offload and VRSS patches), we are now able
to almost saturate a 10G interface between Linux VMs on Hyper-V
on different hosts - close to  9 Gbps as measured via iperf.

Signed-off-by: K. Y. Srinivasan <kys@microsoft.com>
Reviewed-by: Haiyang Zhang <haiyangz@microsoft.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit c25aaf814a63f9d9c4e45416f13d70ef0aa0be2e)
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>