[crypto] qat: Enforce valid numa configuration
Message-id: <1413380663-25484-3-git-send-email-nikolay@redhat.com>
Patchwork-id: 97575
O-Subject: [RHEL7.1 PATCH BZ 1152642 2/2] crypto: qat - Enforce valid numa configuration
Bugzilla: 1152642
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Tadeusz Struk <tadeusz.struk@intel.com>

In a system with NUMA configuration we want to enforce that the accelerator is
connected to a node with memory to avoid cross QPI memory transaction.
Otherwise there is no point in using the accelerator as the encryption in
software will be faster.

Signed-off-by: Tadeusz Struk <tadeusz.struk@intel.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>