[kernel] printk: bump LOG_BUF_SHIFT
Message-id: <x49siied3pf.fsf@segfault.boston.devel.redhat.com>
Patchwork-id: 98329
O-Subject: [rhel7 patch] printk: bump LOG_BUF_SHIFT
Bugzilla: 1156189
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

Hi,

Sysrq-t output in our RHEL 7 kernel is just immense (maybe due to
CONFIG_SCHED_DEBUG?).  Go ahead, try it!  Even with few processes
running, I'm seeing overflows of the ring buffer, which makes debugging
problems in this way difficult (you need to reboot the system, passing a
new log_buf_len= parameter, which is not a great option if you have a
hard to hit hang).

So, either we can reduce the amount of logging, or we can bump the
printk buffer.  I've elected to bump the size of the printk buffer.

This resolves bug 1156189.

Signed-off-by: Jeff Moyer <jmoyer@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>