[alsa] Revert: Kconfig: rename HAS_IOPORT to HAS_IOPORT_MAP
Message-id: <1416436120-55539-1-git-send-email-jarod@redhat.com>
Patchwork-id: 100380
O-Subject: [RHEL7 PATCH] alsa: Revert: Kconfig: rename HAS_IOPORT to HAS_IOPORT_MAP
Bugzilla: 1112200
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>
RH-Acked-by: Jaroslav Kysela <jkysela@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>

Bugzilla: 1112200

This reverts commit 8fbe6eb026ed37a8154dea15d55ab728a01f6f36.

This was a partial cherry-pick of an upstream change that also changed the
system-wide HAS_IOPORT to HAS_IOPORT_MAP. The problem with that is that
without that system-wide part of the change, an alsa sound driver we were
building in RHEL7.0 is now dependent on a config var that doesn't actually
exist in our tree. So lets just back this out and let lx6464es.ko build
again.

CC: Jaroslav Kysela <jkysela@redhat.com>
CC: Ivan Vecera <ivecera@redhat.com>
CC: Prarit Bhargava <prarit@redhat.com>
CC: John Linville <linville@redhat.com>
CC: Paul Bunyan <pbunyan@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>