[net] bonding: remove bond_vlan_used()
Message-id: <1416574342-24493-40-git-send-email-nikolay@redhat.com>
Patchwork-id: 100457
O-Subject: [RHEL7.2 PATCH BZ 1166647 039/336] bonding: remove bond_vlan_used()
Bugzilla: 1166647
RH-Acked-by: Neil Horman <nhorman@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Jiri Pirko <jpirko@redhat.com>

From: Veaceslav Falico <vfalico@redhat.com>

We're using it currently to verify if we have vlans before getting the tag
from the skb we're about to send. It's useless because the vlan_get_tag()
verifies if the skb has the tag (and returns an error if not), and we can
receive tagged skbs only if we *already* have vlans.

Plus, the current RCUed implementation is kind of useless anyway - the we
can remove the last vlan in the moment we return from the function.

So remove the only usage of it and the whole function.

CC: Jay Vosburgh <fubar@us.ibm.com>
CC: Andy Gospodarek <andy@greyhouse.net>
Signed-off-by: Veaceslav Falico <vfalico@redhat.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 6f477d420183f7cacc73f46a93e12961cb234e91)
Signed-off-by: Rafael Aquini <aquini@redhat.com>

Conflicts:
	drivers/net/bonding/bonding.h
	- Due to our use of upper dev lists, anyhow removed the
	  function.

Signed-off-by: Jiri Benc <jbenc@redhat.com>